Size is not grandeur

+++ 0x00-hello_world/6-size.c
@@ -5,8 +5,8 @@
  */
 int main(void)
 {
-	printf("Size of a char: %i byte(s)\n", sizeof(char));
-	printf("Size of an int: %i byte(s)\n", sizeof(long));
-	printf("Size of a long int: %i byte(s)\n", sizeof(long long int));
-	printf("Size of a float: %i byte(s)\n", sizeof(float));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long) sizeof(char));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(long));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(long long int));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(float));
 	return (0);
